refactor(reviewer-bot): extract command handlers
